#1486 Implementar mètodes de gestió i consulta de dades

+++ microservices/helium-dada-service/src/main/java/es/caib/helium/dada/service/ExpedientServiceImpl.java
@@ -1,15 +1,19 @@
 package es.caib.helium.dada.service;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import es.caib.helium.dada.domain.Dada;
-import es.caib.helium.dada.domain.DadaMS;
-import es.caib.helium.dada.domain.ExpedientCapcalera;
+import es.caib.helium.dada.domain.Expedient;
+import es.caib.helium.dada.model.Filtre;
+import es.caib.helium.dada.model.FiltreCapcalera;
+import es.caib.helium.dada.model.Ordre;
 import es.caib.helium.dada.model.PagedList;
 import es.caib.helium.dada.repository.DadaRepository;
 import es.caib.helium.dada.repository.ExpedientRepository;
@@ -24,20 +28,29 @@ public class ExpedientServiceImpl implements ExpedientService {
 
 	private final ExpedientRepository expedientRepository;
 	private final DadaRepository dadaRepository;
-	
+
 	// Dades capçalera expedient
 
 	@Override
-	public PagedList<Dada> consultaResultats(int page, int size) {
-		log.debug("[SRV] Consultant resultats");
-		var pageable = PageRequest.of(page, size);	
-		var pagina = dadaRepository.findAll(pageable); 
-		var content = pagina.getContent();
-		return new PagedList<Dada>(pagina.getContent(), pageable, pagina.getContent().size());
+	public PagedList<Expedient> consultaResultats(Map<String, Filtre> filtreValors, Integer entornId,
+			Integer expedientTipusId, List<Ordre> ordre, Integer page, Integer size) {
+
+		log.debug("[SRV] Consultant resultats paginats");
+		var pagina = expedientRepository.findByFiltres(filtreValors, entornId, expedientTipusId, ordre, page, size);
+		var pageable = PageRequest.of(page, pagina.size() > 0 ? pagina.size() : 1);
+		return new PagedList<Expedient>(pagina, pageable, pagina.size());
 	}
-	
+
 	@Override
-	public void createExpedient(ExpedientCapcalera expedient) {
+	public List<Expedient> consultaResultats(Map<String, Filtre> filtreValors, Integer entornId,
+			Integer expedientTipusId, List<Ordre> ordre) {
+
+		log.debug("[SRV] Consultant resultats");
+		return expedientRepository.findByFiltres(filtreValors, entornId, expedientTipusId, ordre, null, null);
+	}
+
+	@Override
+	public void createExpedient(Expedient expedient) {
 
 		log.debug("[SRV] Creant dades capçalera expedient (expedient= " + expedient.toString());
 		var exp = expedientRepository.findByExpedientId(expedient.getExpedientId());
@@ -49,7 +62,23 @@ public class ExpedientServiceImpl implements ExpedientService {
 	}
 
 	@Override
-	public ExpedientCapcalera findByExpedientId(Long expedientId) {
+	public void createExpedients(List<Expedient> expedients) {
+
+		log.debug("[SRV] Creant dades capçalera expedient per múltiples expedients");
+		List<Expedient> exps = new ArrayList<>();
+		for (var expedient : expedients) {
+			if (expedientRepository.findByExpedientId(expedient.getExpedientId()).isPresent()) {
+				log.debug(
+						"[SRV] Expedient (expedientId=" + expedient.getExpedientId() + " ja existeix, no es guardarà");
+				continue;
+			}
+			exps.add(expedient);
+		}
+		expedientRepository.saveAll(exps);
+	}
+
+	@Override
+	public Expedient findByExpedientId(Long expedientId) {
 
 		log.debug("[SRV] Buscant les dades de capçalera de l'expedient (expedientId= " + expedientId);
 		var expedient = expedientRepository.findByExpedientId(expedientId);
@@ -61,31 +90,77 @@ public class ExpedientServiceImpl implements ExpedientService {
 	}
 
 	@Override
-	public void deleteExpedient(Long expedientId) {
+	public boolean deleteExpedient(Long expedientId) {
 
 		log.debug("[SRV] Borrant dades capçalera expedient (expedientId= " + expedientId);
+		var expedient = findByExpedientId(expedientId);
+		if (expedient == null) {
+			return false; // Aixi es pot retornar el 404 al Controller
+		}
 		expedientRepository.deleteByExpedientId(expedientId);
+		//TODO falta esborrar els documents de la coleccio dades que fan referencia a aquest expedient. 
+		return true;
 	}
 
 	@Override
-	public void putExpedient(Long expedientId, ExpedientCapcalera expedient) {
+	public void deleteExpedients(List<Long> expedients) {
 
-		var exp = findByExpedientId(expedientId);
-		exp.setEntornId(expedient.getEntornId());
-		exp.setTipusId(expedient.getTipusId());
-		exp.setNumero(expedient.getNumero());
-		exp.setTitol(expedient.getTitol());
-		exp.setProcesPrincipalId(expedient.getProcesPrincipalId());
-		exp.setEstatId(expedient.getEstatId());
-		exp.setDataInici(expedient.getDataInici());
-		exp.setDataFi(expedient.getDataFi());
-		expedientRepository.save(exp);
+		log.debug("[SRV] Borrant dades capçalera múltiples expedients");
+		var exps = expedientRepository.findByExpedients(expedients);
+		if (exps.isPresent()) {
+			expedientRepository.deleteAll(exps.get());
+			//TODO falta esborrar els documents de la coleccio dades que fan referencia a aquests expedient. 
+		}
 	}
 
 	@Override
-	public void patchExpedient(Long expedientId, ExpedientCapcalera expedient) {
+	public boolean putExpedient(Long expedientId, Expedient expedient) {
+
+		// TODO CANVIAR LES CRIDES QUE FALATEN AL REPOSITORI EN COMPTES DE LA FUNCIÓ
+		var exp = expedientRepository.findByExpedientId(expedientId);
+		if (exp.isEmpty()) {
+			return false;
+		}
+		exp.get().setEntornId(expedient.getEntornId());
+		exp.get().setTipusId(expedient.getTipusId());
+		exp.get().setNumero(expedient.getNumero());
+		exp.get().setTitol(expedient.getTitol());
+		exp.get().setProcesPrincipalId(expedient.getProcesPrincipalId());
+		exp.get().setEstatId(expedient.getEstatId());
+		exp.get().setDataInici(expedient.getDataInici());
+		exp.get().setDataFi(expedient.getDataFi());
+		expedientRepository.save(exp.get());
+		return true;
+	}
+
+	@Override
+	public void putExpedients(List<Expedient> expedients) {
+
+		List<Expedient> exps = new ArrayList<>();
+		for (var expedient : expedients) {
+			var exp = findByExpedientId(expedient.getExpedientId());
+			if (exp != null) {
+				exp.setEntornId(expedient.getEntornId());
+				exp.setTipusId(expedient.getTipusId());
+				exp.setNumero(expedient.getNumero());
+				exp.setTitol(expedient.getTitol());
+				exp.setProcesPrincipalId(expedient.getProcesPrincipalId());
+				exp.setEstatId(expedient.getEstatId());
+				exp.setDataInici(expedient.getDataInici());
+				exp.setDataFi(expedient.getDataFi());
+				exps.add(exp);
+			}
+		}
+		expedientRepository.saveAll(exps);
+	}
+
+	@Override
+	public boolean patchExpedient(Long expedientId, Expedient expedient) {
 
 		var exp = findByExpedientId(expedientId);
+		if (exp == null) {
+			return false;
+		}
 		exp.setEntornId(expedient.getEntornId() != null ? expedient.getEntornId() : exp.getEntornId());
 		exp.setTipusId(expedient.getTipusId() != null ? expedient.getTipusId() : exp.getTipusId());
 		exp.setNumero(expedient.getNumero() != null && !expedient.getNumero().isEmpty() ? expedient.getNumero()
@@ -98,6 +173,31 @@ public class ExpedientServiceImpl implements ExpedientService {
 		exp.setDataInici(expedient.getDataInici() != null ? expedient.getDataInici() : exp.getDataInici());
 		exp.setDataFi(expedient.getDataFi() != null ? expedient.getDataFi() : exp.getDataFi());
 		expedientRepository.save(exp);
+		return true;
+	}
+
+	@Override
+	public void patchExpedients(List<Expedient> expedients) {
+
+		List<Expedient> exps = new ArrayList<>();
+		for (var expedient : expedients) {
+			var exp = findByExpedientId(expedient.getExpedientId());
+			if (exp != null) {
+				exp.setEntornId(expedient.getEntornId() != null ? expedient.getEntornId() : exp.getEntornId());
+				exp.setTipusId(expedient.getTipusId() != null ? expedient.getTipusId() : exp.getTipusId());
+				exp.setNumero(expedient.getNumero() != null && !expedient.getNumero().isEmpty() ? expedient.getNumero()
+						: exp.getNumero());
+				exp.setTitol(expedient.getTitol() != null && !expedient.getTitol().isEmpty() ? expedient.getTitol()
+						: exp.getTitol());
+				exp.setProcesPrincipalId(expedient.getProcesPrincipalId() != null ? expedient.getProcesPrincipalId()
+						: exp.getProcesPrincipalId());
+				exp.setEstatId(expedient.getEstatId() != null ? expedient.getEstatId() : exp.getEstatId());
+				exp.setDataInici(expedient.getDataInici() != null ? expedient.getDataInici() : exp.getDataInici());
+				exp.setDataFi(expedient.getDataFi() != null ? expedient.getDataFi() : exp.getDataFi());
+				exps.add(exp);
+			}
+		}
+		expedientRepository.saveAll(exps);
 	}
 
 	// Dades expedient
@@ -136,12 +236,11 @@ public class ExpedientServiceImpl implements ExpedientService {
 		log.debug("[SRV] Creant dades per l'expedient " + expedientId + " amb procesId " + procesId);
 
 		for (var dada : dades) {
-			// TODO VALIDAR SI JA EXISTEIX?¿
-			
+
 			var d = dadaRepository.findByExpedientIdAndProcesIdAndCodi(expedientId, procesId, dada.getCodi());
 			if (d.isPresent()) {
-				// Evitar dades amb el codi repetit
-				log.error("[ExpedientServiceImpl.crearDada] Expedient amb id " + expedientId + " procesId " + procesId 
+				// Evitar dades amb el codi repetit segons proces i codi
+				log.error("[ExpedientServiceImpl.crearDada] Expedient amb id " + expedientId + " procesId " + procesId
 						+ " ja té una dada amb codi " + dada.getCodi());
 				continue;
 			}
@@ -152,12 +251,11 @@ public class ExpedientServiceImpl implements ExpedientService {
 	}
 
 	@Override
-	public void putDadaByExpedientIdAndCodi(Long expedientId, String codi, Dada dada) {
+	public boolean putDadaByExpedientIdAndCodi(Long expedientId, String codi, Dada dada) {
+		
 		var dadaOptional = dadaRepository.findByExpedientIdAndCodi(expedientId, codi);
 		if (dadaOptional.isEmpty()) {
-			log.error("[ExpedientServiceImpl.putDadaByExpedientIdAndCodi] No existeix la dada amb expedientId  "
-					+ expedientId + " i codi " + codi);
-			return;
+			return false;
 		}
 
 		var dadaMongo = dadaOptional.get();
@@ -165,18 +263,19 @@ public class ExpedientServiceImpl implements ExpedientService {
 		dadaMongo.setTipus(dada.getTipus());
 		dadaMongo.setValor(dada.getValor());
 		dadaRepository.save(dadaMongo);
+		return true;
 	}
 
 	@Override
-	public void deleteDadaByExpedientIdAndCodi(Long expedientId, String codi) {
-		// TODO PREGUNTAR IS ÉS UN DELETE ALL PERQUÈ SINO AIXÒ POT RETORNAR MÉS D'UNA DADA, O NO, DEPEN DE LES VALIDACIONS
+	public boolean deleteDadaByExpedientIdAndCodi(Long expedientId, String codi) {
+		// TODO PREGUNTAR IS ÉS UN DELETE ALL PERQUÈ SINO AIXÒ POT RETORNAR MÉS D'UNA
+		// DADA, O NO, DEPEN DE LES VALIDACIONS
 		var dada = dadaRepository.findByExpedientIdAndCodi(expedientId, codi);
 		if (dada.isEmpty()) {
-			log.error("[ExpedientServiceImpl.deleteByExpedientIdAndCodi] No existeix la dada amb expedientId "
-					+ expedientId + " i codi " + codi);
-			return;
+			return false;
 		}
 		dadaRepository.delete(dada.get());
+		return true;
 	}
 
 	@Override
@@ -184,7 +283,8 @@ public class ExpedientServiceImpl implements ExpedientService {
 
 		var dadesMongo = new ArrayList<Dada>();
 		for (var dada : dades) {
-			if (dada.getCodi() == null || dada.getCodi().isEmpty()) { // TODO MIRAR QUE NO ESTIGUI EL CODI REPETIT A LA VARIABLE dades
+			if (dada.getCodi() == null || dada.getCodi().isEmpty()) { // TODO MIRAR QUE NO ESTIGUI EL CODI REPETIT A LA
+																		// VARIABLE dades
 				continue;
 			}
 			var dadaMongo = dadaRepository.findByExpedientIdAndProcesIdAndCodi(expedientId, procesId, dada.getCodi());
@@ -204,30 +304,30 @@ public class ExpedientServiceImpl implements ExpedientService {
 	}
 
 	@Override
-	public void putDadaByExpedientIdProcesIdAndCodi(Long expedientId, Long procesId, String codi, Dada dada) {
+	public boolean putDadaByExpedientIdProcesIdAndCodi(Long expedientId, Long procesId, String codi, Dada dada) {
 
 		var dadaOptional = dadaRepository.findByExpedientIdAndProcesIdAndCodi(expedientId, procesId, codi);
 		if (dadaOptional.isEmpty()) {
-			log.error("[ExpedientServiceImpl.deleteByExpedientIdAndCodi] No existeix la dada amb expedientId "
-					+ expedientId + " i codi " + codi);
-			return;
+			return false;
 		}
 		var dadaMongo = dadaOptional.get();
 		dadaMongo.setMultiple(dada.isMultiple());
 		dadaMongo.setTipus(dada.getTipus());
 		dadaMongo.setValor(dada.getValor());
 		dadaRepository.save(dadaMongo);
+		return true;
 	}
 
 	@Override
-	public void deleteDadaByExpedientIdAndProcesIdAndCodi(Long expedientId, Long procesId, String codi) {
-		
+	public boolean deleteDadaByExpedientIdAndProcesIdAndCodi(Long expedientId, Long procesId, String codi) {
+
 		var dada = dadaRepository.findByExpedientIdAndProcesIdAndCodi(expedientId, procesId, codi);
 		if (dada.isEmpty()) {
 			log.error("[ExpedientServiceImpl.deleteByExpedientIdAndCodi] No existeix la dada amb expedientId "
 					+ expedientId + " i codi " + codi);
-			return;
+			return false;
 		}
 		dadaRepository.delete(dada.get());
+		return true;
 	}
 }
